+++ packages/typescript/src/index.ts
@@ -1,3 +1,4 @@
 export * from './languageService';
 export * from './languageServiceHost';
 export * from './sys';
+export * from './dtsHost';
