+++ packages/typescript/tsconfig.build.json
@@ -15,8 +15,5 @@
 		{
 			"path": "../language-service/tsconfig.build.json"
 		},
-		{
-			"path": "../web-fs/tsconfig.build.json"
-		},
 	],
 }
