--- packages/web-fs/package.json
@@ -1,18 +1,0 @@
-{
-	"name": "@volar/web-fs",
-	"version": "1.6.9",
-	"main": "out/index.js",
-	"license": "MIT",
-	"files": [
-		"out/**/*.js",
-		"out/**/*.d.ts"
-	],
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/volarjs/volar.js.git",
-		"directory": "packages/web-fs"
-	},
-	"devDependencies": {
-		"@volar/language-service": "^1.6.9"
-	}
-}
