--- packages/web-fs/src/index.ts
@@ -1,1 +1,0 @@
-export * from './dtsHost';
