--- packages/web-fs/tsconfig.build.json
@@ -1,16 +1,0 @@
-{
-	"extends": "../../tsconfig.json",
-	"compilerOptions": {
-		"noEmit": false,
-		"outDir": "out",
-		"rootDir": "src",
-	},
-	"include": [
-		"src"
-	],
-	"references": [
-		{
-			"path": "../language-service/tsconfig.build.json"
-		},
-	],
-}
